requirements.txt
pycryptodome==3.20.0
python-dotenv==1.0.0
psutil==5.9.6
colorlog==6.7.0
typing-extensions==4.8.0
scrypt==0.8.22

